refactor: utils magic and town
