+++ app/utils/magicUtils.ts
