+++ app/utils/townUtils.ts
